Norminette ok on parser.c
