+++ Makefile
@@ -6,7 +6,7 @@
 #    By: lfarias- <lfarias-@student.42.rio>         +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/12/27 14:14:19 by lfarias-          #+#    #+#              #
-#    Updated: 2023/01/06 14:48:50 by lfarias-         ###   ########.fr        #
+#    Updated: 2023/01/06 15:08:06 by lfarias-         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -25,7 +25,7 @@ LDFLAGS 	=   -L  /Users/lfarias-/.brew/Cellar/readline/8.2.1/lib
 SRC			= 	main.c command_executor.c command_loader.c error_handler.c \
 				signal_handlers.c echo.c exit.c build_env.c env.c pwd.c cd.c \
 				export.c expander.c kludge.c unset.c quote_resolver.c \
-				lexer.c parser.c interpreter.c \
+				lexer.c parser.c parser_rules.c interpreter.c \
 
 SRCS		= 	$(addprefix src/,$(SRC))
